% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-selenium.R
\name{click_element}
\alias{click_element}
\title{Click the element}
\usage{
click_element(element)
}
\arguments{
\item{element}{HTML element.}
}
\value{
Returns the status of the action.
}
\description{
Click the element
}
\keyword{internal}
