% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-digimap.R
\name{digimap_os}
\alias{digimap_os}
\title{(Assisted) request of EDINA's Digimap data from the Ordnance Survey}
\source{
https://digimap.edina.ac.uk/os
}
\usage{
digimap_os(
  client,
  area_name = NULL,
  dataset = NULL,
  format = NULL,
  version = NULL,
  org = Sys.getenv("ORG"),
  sleep = 1
)
}
\arguments{
\item{client}{\code{RSelenium} client.}

\item{area_name}{String with UK national grid name (e.g., 'SD', 'SD20'). See
\url{ordnancesurvey.co.uk/documents/resources/guide-to-nationalgrid.pdf}.}

\item{dataset}{String with the name of the data set to download (e.g.,
'NTM' for National Tree Map or 'Terrain-5 DTM' for the OS Terrain 5
Digital Terrain Model). See
\url{https://digimap.edina.ac.uk/help/our-maps-and-data/os_products/}.}

\item{format}{String with the data format (e.g., 'SHAPE'). See
\url{https://digimap.edina.ac.uk/help/our-maps-and-data/os_products/}.}

\item{version}{String with the version of the data set (e.g., 'July 2023').
See
\url{https://digimap.edina.ac.uk/help/our-maps-and-data/os_products/}.}

\item{org}{String with your organisation name (for login purposes, this
can be done manually). Done only once per session.}

\item{sleep}{Integer with number of seconds to use as pause between actions
on the web page.}
}
\value{
Logic value with the status of the data request.
}
\description{
Digimap’s Ordnance Survey collection provides a full range of topographic
Ordnance Survey data for the UK.
Note that this function helps you to request the data, once your request
has been processed, you will have to manually download the data (email
instructions will be provided by Digimap).
}
