% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-datetime.R
\name{duration2datetime}
\alias{duration2datetime}
\title{Convert duration to date-time
Convert a string with a duration (e.g. 'an hour ago') to a date-time string,
based on a reference time}
\usage{
duration2datetime(
  str,
  ref_time = Sys.time(),
  output_format = "\%Y-\%m-\%d \%H:\%M:\%S \%Z"
)
}
\arguments{
\item{str}{String with a duration (see examples)}

\item{ref_time}{Reference time (default: \code{Sys.time()}, current time)}

\item{output_format}{String with the format of the output (default:
\code{"\%Y-\%m-\%d \%H:\%M:\%S \%Z"})}
}
\value{
Date-time object based on the input string, \code{str}, and the reference
time \code{ref_time}.
}
\description{
Convert duration to date-time
Convert a string with a duration (e.g. 'an hour ago') to a date-time string,
based on a reference time
}
\examples{
duration2datetime("a day ago")
duration2datetime("an hour ago")
duration2datetime("2 days ago")
duration2datetime("2 hours ago")
}
