% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-selenium.R
\name{execute_script}
\alias{execute_script}
\title{Inject a snippet of JavaScript}
\usage{
execute_script(client, script, args)
}
\arguments{
\item{client}{\code{RSelenium} client.}

\item{script}{String with the JavaScript to be executed.}

\item{args}{List of srguments for the \code{script}.}
}
\value{
Returns the status of the action.
}
\description{
Inject a snippet of JavaScript into the page for exe- cution in the context
of the currently selected frame. The executed script is assumed to be
synchronous and the result of evaluating the script is returned to the
client. The script argu- ment defines the script to execute in the form of a
function body.
}
\keyword{internal}
