% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape-google.R
\name{expand_reviews}
\alias{expand_reviews}
\title{Expand reviews}
\usage{
expand_reviews(
  client,
  using = "xpath",
  value = "//button[@jsaction='pane.review.expandReview']"
)
}
\arguments{
\item{client}{\code{RSelenium} client.}

\item{using}{String with "css" or "xpath".}

\item{value}{String with css tag or xpath.}
}
\description{
Expand long reviews to capture all the comment
}
\keyword{internal}
