% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape-google.R
\name{expand_reviews_2}
\alias{expand_reviews_2}
\title{Expand reviews}
\usage{
expand_reviews_2(
  element,
  using = "xpath",
  value = "//button[contains(@jsaction, 'expandReview')]"
)
}
\arguments{
\item{element}{Reference element.}

\item{using}{String with "css" or "xpath".}

\item{value}{String with css tag or xpath.}
}
\description{
Expand long reviews to capture all the comment
}
\keyword{internal}
