% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-nhs.R
\name{find_a_gp}
\alias{find_a_gp}
\title{Scrape GP practices}
\usage{
find_a_gp(
  client,
  postcode,
  base = "https://www.nhs.uk/service-search/find-a-gp",
  sleep = 1
)
}
\arguments{
\item{client}{\code{RSelenium} client.}

\item{postcode}{String with the target postcode.}

\item{base}{String with the base URL for Google Maps website.}

\item{sleep}{Integer with number of seconds to use as pause between actions
on the web page.}
}
\value{
Data frame with GP practices near the given \code{postcode}.
}
\description{
Scrape GP practices near a given postcode
}
\examples{
\dontrun{
# Create RSelenium session
rD <- RSelenium::rsDriver(browser = "firefox", port = 4544L, verbose = FALSE)

# Retrieve GP practices near L69 3GL
# (Waterhouse building, University of Liverpool)
wh_gps_tb <- scrappy::find_a_gp(rD$client, postcode = "L69 3GL")

# Stop server
rD$server$stop()
}
}
