% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-selenium.R
\name{find_and_click}
\alias{find_and_click}
\title{Search for an element on the page and click it}
\usage{
find_and_click(client, using, value)
}
\arguments{
\item{client}{\code{RSelenium} client.}

\item{using}{String with "css" or "xpath".}

\item{value}{String with css tag or xpath.}
}
\value{
The located element will be returned as an object of \code{webElement}
class.
}
\description{
Search for an element on the page, starting from the document root.
}
\keyword{internal}
