% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-selenium.R
\name{find_child_element}
\alias{find_child_element}
\title{Search for a child element for current element}
\usage{
find_child_element(element, using, value)
}
\arguments{
\item{element}{Reference element.}

\item{using}{String with "css" or "xpath".}

\item{value}{String with css tag or xpath.}
}
\value{
The located element will be returned as an object of \code{webElement}
class.
}
\description{
Search for a child element for current element
}
\keyword{internal}
