% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape-google.R
\name{handle_cookies}
\alias{handle_cookies}
\title{Handle cookies}
\usage{
handle_cookies(
  client,
  using = "xpath",
  value = "//button[@aria-label='Reject all']",
  accept_cookies = FALSE,
  sleep = 1
)
}
\arguments{
\item{client}{\code{RSelenium} client.}

\item{using}{String with "css" or "xpath".}

\item{value}{String with css tag or xpath.}

\item{accept_cookies}{Not used (to be implemented).}

\item{sleep}{Integer with number of seconds to use as pause between actions
on the web page.}
}
\description{
Handle cookies page
}
\keyword{internal}
