% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-selenium.R
\name{navigate}
\alias{navigate}
\title{Navigate to a given url}
\usage{
navigate(client, url)
}
\arguments{
\item{client}{\code{RSelenium} client.}

\item{url}{String with a URL.}
}
\value{
Returns the status of the action.
}
\description{
Navigate to a given url
}
\keyword{internal}
