% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{newa3_stations}
\alias{newa3_stations}
\title{NEWA v3 Weather Stations dataset}
\format{
A data frame with 801 rows and 10 variables:
\describe{
\item{name}{Station's name.}
\item{state}{State where the station is located.}
\item{code}{Station's code.}
\item{affiliation}{Entity to which the entity is affiliated.}
\item{affiliation_url}{Entity's URL.}
\item{latitude}{Station's latitude.}
\item{longitude}{Station's longitude.}
\item{elevation}{Station's elevation.}
\item{start_year}{Start year (data available).}
\item{is_icao}{Boolean flag to indicate if the station is part of the
International Civil Aviation Organization (ICAO) (e.g is an airport).}
}
}
\source{
\url{https://newa.cornell.edu}
}
\usage{
data(newa3_stations)
}
\description{
A dataset containing information of 801 weather stations in the Network for
Environment and Weather Applications (NEWA) version 3 at Cornell University.
}
\author{
Network for Environment and Weather Applications
\email{support@newa.zendesk.com}
}
\keyword{datasets}
