% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrappy.R
\name{newa_nrcc}
\alias{newa_nrcc}
\title{Retrieve data from NEWA at Cornell University}
\usage{
newa_nrcc(
  client,
  year,
  month,
  station,
  base = "http://newa.nrcc.cornell.edu/newaLister",
  interval = "hly",
  sleep = 6
)
}
\arguments{
\item{client}{\code{RSelenium} client.}

\item{year}{Numeric value with the year.}

\item{month}{Numeric value with the month.}

\item{station}{String with the station abbreviation. Check the
\url{http://newa.cornell.edu/index.php?page=station-pages} for a list.}

\item{base}{Base URL
(default: \url{http://newa.nrcc.cornell.edu/newaLister}).}

\item{interval}{String with data interval (default: hly, hourly).}

\item{sleep}{Numeric value with the number of seconds to wait for the page
to load the results (default: 6 seconds).}
}
\description{
Retrieve Weather data from the Network for Environment and Weather
Applications (NEWA) at Cornell University.
}
\examples{
\dontrun{
scrappy::newa_nrcc(my_client, 2020, 12, "gbe")
}
}
