% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrappy.R
\name{newa_nrcc3}
\alias{newa_nrcc3}
\title{Retrieve data from NEWA v3.0 at Cornell University}
\usage{
newa_nrcc3(
  year,
  month,
  day,
  hour,
  station,
  base = "https://hrly.nrcc.cornell.edu/stnHrly"
)
}
\arguments{
\item{year}{Numeric value with the start year.}

\item{month}{Numeric value with the start month.}

\item{day}{Numeric value with the start day.}

\item{hour}{Numeric value with the start hour.}

\item{station}{String with the station abbreviation. Check
\code{scrappy::newa3_stations} for a list of stations and abbreviations.}

\item{base}{Base URL
(default: \url{https://hrly.nrcc.cornell.edu/stnHrly}).}
}
\value{
List of data frames with \code{hourly}, \code{daily},
\code{hourly_forecast}, and daily forecast (\code{daily_forecast}) data.
}
\description{
Retrieve Weather data from the Network for Environment and Weather
Applications (NEWA) version 3.0 at Cornell University.
}
\examples{
scrappy::newa_nrcc3(2021, 12, 01, 00, "gbe")
}
