% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{newa_stations}
\alias{newa_stations}
\title{NEWA Weather Stations dataset}
\format{
A data frame with 718 rows and 3 variables:
\describe{
\item{name}{Station name.}
\item{state}{State name.}
\item{code}{Station code.}
}
}
\source{
\url{http://newa.cornell.edu/index.php?page=station-pages}
}
\usage{
data(newa_stations)
}
\description{
A dataset containing information of 718 weather stations in the Network for
Environment and Weather Applications (NEWA) at Cornell University.
}
\author{
Network for Environment and Weather Applications
\email{support@newa.zendesk.com}
}
\keyword{datasets}
