% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape-google.R
\name{open_overview_tab}
\alias{open_overview_tab}
\title{Open overview tab}
\usage{
open_overview_tab(
  client,
  using = "xpath",
  value = "//button[@role='tab' and contains(., 'Overview')]",
  sleep = 1
)
}
\arguments{
\item{client}{\code{RSelenium} client.}

\item{using}{String with "css" or "xpath".}

\item{value}{String with css tag or xpath.}

\item{sleep}{Integer with number of seconds to use as pause between actions
on the web page.}
}
\description{
Open overview tab
}
\keyword{internal}
