% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-nhs.R
\name{parse_gpp}
\alias{parse_gpp}
\title{Parse list of GP practices}
\usage{
parse_gpp(gpp)
}
\arguments{
\item{gpp}{List of GP practices (\code{webElement} class)}
}
\value{
Tibble with the GP practices
}
\description{
Parse list of GP practices
}
\keyword{internal}
