% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape-google.R
\name{parse_reviews}
\alias{parse_reviews}
\title{Parse list of reviews}
\usage{
parse_reviews(reviews)
}
\arguments{
\item{reviews}{List of reviews (\code{webElement} class)}
}
\value{
Tibble with the reviews
}
\description{
Parse list of reviews
}
\keyword{internal}
