% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotScrappy.R
\name{plotScrappy}
\alias{plotScrappy}
\title{Make a scrappy plot}
\usage{
plotScrappy(object, dimred, great_qual = 5, good_qual = 10, ok_qual = 15)
}
\arguments{
\item{object}{Single cell object. Scrappy recognizes SingleCellExperiment objects or Seurat objects}

\item{dimred}{dimension reductions object to be accessed}

\item{great_qual}{Percentage of mitochondrial gene expression for a cell to be considered great quality (smaller than)}

\item{good_qual}{Percentage of mitochondrial gene expression for a cell to be considered good quality (smaller than)}

\item{ok_qual}{Percentage of mitochondrial gene expression for a cell to be considered ok quality (smaller than). Any other percentage above is considered poop quality}
}
\value{

}
\description{
Make a scrappy plot
}
\examples{

}
