% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-selenium.R
\name{random_wait}
\alias{random_wait}
\title{Random wait time}
\usage{
random_wait(sleep = 1)
}
\arguments{
\item{sleep}{Integer with number of seconds to use as pause between actions
on the web page.}
}
\description{
Random wait time
}
\keyword{internal}
