% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrappy.R
\docType{package}
\name{scrappy-package}
\alias{scrappy}
\alias{scrappy-package}
\title{scrappy: A Simple Web Scraper}
\description{
A group of functions to scrape data from different websites, for academic purposes.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/villegar/scrappy/}
  \item \url{https://villegar.github.io/scrappy/}
  \item \url{https://scrappy.robertovillegas-diaz.com/}
  \item Report bugs at \url{https://github.com/villegar/scrappy/issues/}
}

}
\author{
\strong{Maintainer}: Roberto Villegas-Diaz \email{r.villegas-diaz@outlook.com} (\href{https://orcid.org/0000-0001-5036-8661}{ORCID})

Authors:
\itemize{
  \item Roberto Villegas-Diaz \email{r.villegas-diaz@outlook.com} (\href{https://orcid.org/0000-0001-5036-8661}{ORCID})
}

}
\keyword{internal}
