% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape-google.R
\name{scroll_reviews}
\alias{scroll_reviews}
\title{Scroll reviews}
\usage{
scroll_reviews(
  client,
  using = "css",
  value = "div.m6QErb.DxyBCb.kA9KIf.dS8AEf",
  scroll_px = 1000,
  sleep = 1
)
}
\arguments{
\item{client}{\code{RSelenium} client.}

\item{using}{String with "css" or "xpath".}

\item{value}{String with css tag or xpath.}

\item{scroll_px}{Integer with the number of pixels to scroll.}

\item{sleep}{Integer with number of seconds to use as pause between actions
on the web page.}
}
\description{
Scroll reviews container and load more reviews
}
\keyword{internal}
