% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-web.R
\name{wait_to_load}
\alias{wait_to_load}
\title{Wait until page has finished loading}
\usage{
wait_to_load(client, using = "css", value = "body", sleep = 1)
}
\arguments{
\item{client}{\code{RSelenium} client.}

\item{using}{String with the property to use to find the element (e.g. "css",
"xpath", etc.) (default: "css").}

\item{value}{String with the tag of the page element to wait to load
(default: "body").}

\item{sleep}{Numeric value with the number of seconds to wait for the page
to load the results (default: 1 second).}
}
\description{
Wait until page has finished loading the element with the tag \code{value}
}
