% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape-google.R
\name{overall_rating}
\alias{overall_rating}
\title{Overall rating of the place}
\usage{
overall_rating(client)
}
\arguments{
\item{client}{\code{RSelenium} client.}
}
\value{
Tibble with average number of stars and total reviews
}
\description{
Overall rating of the place
}
\keyword{internal}
